ethdev: add rte_device to port_id function
Some EAL facilities are using rte_devices. Once the rte_device of a
port is found, some need to transform it into a port id to be able to
use the ethdev API.

Add a function to transform an rte_device pointer into an ethdev port id,
as well as an ownership aware variant.

Use this new API in a few PMDs that could be simplified with it.

Cc: Chas Williams <chas3@att.com>
Cc: Liron Himi <lironh@marvell.com>
Cc: Andrew Rybchenko <arybchenko@solarflare.com>
Cc: Thomas Monjalon <thomas@monjalon.net>

Signed-off-by: Gaetan Rivet <grive@u256.net>